factor out grpc, ws and registry servers
